Rename getLatestSlic3rRelease.sh to getLatestPrusaSlicerRelease.sh
This adds documentation to getLatestPrusaSlicerRelease.sh about
what all the options do (to see it, run without parameters/args).

Some more attention has been paid to the unused portions of
getLatest (url_ver and name_ver). The expected inputs have been
documented as well.

The name change has been reflected in the Dockerfile where it is
used, and the README file, where it shows up in the documentation.